+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - (***BREAKING***) Conversion of arguments in JS callbacks of Python code is now automatic only if it can be correctly deduced from the context, `toJS()` must be called in all other cases
  - Implement Python type coercion
  - Support Python sets
+ 
 #### Bug Fixes
 - Fix [#25](https://github.com/mmomtchev/pymport/issues/25), type-ahead provokes a crash in Node.js 14.x REPL