+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] WIP
+ - Multithreading safety
+
 ### [1.2.1] WIP
  - Fix [[#27](https://github.com/mmomtchev/pymport/issues/27), a function as a single argument is considered a kwargs object